feat: search convo screen code refactoring and builder
